added address functionality, added cartfuntionality,added WishList functionality, added role create three email on app start up,added admin access inside the product and categories api

+++ src/main/java/com/qalaa/category/controller/CategoryController.java
@@ -2,7 +2,9 @@ package com.qalaa.category.controller;
 
 import com.qalaa.category.model.Category;
 import com.qalaa.category.service.CategoryService;
+import com.qalaa.enums.RoleEnum;
 import com.qalaa.util.ApiResponse;
+import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -14,6 +16,7 @@ import java.util.Optional;
 
 @RestController
 @RequestMapping("/api/categories")
+@SecurityRequirement(name = "bearerAuth")
 public class CategoryController {
 
     @Autowired
@@ -32,18 +35,28 @@ public class CategoryController {
 
     }
     @PostMapping
-    public ResponseEntity<ApiResponse<Category>> createCategory(@RequestBody Category category,HttpServletRequest request) {
+    public ResponseEntity<ApiResponse<Category>> createCategory(@RequestAttribute("id") Long userId,@RequestAttribute("role") String role,@RequestBody Category category,HttpServletRequest request) {
+        if(!role.equals(RoleEnum.ADMIN.name())){
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(new ApiResponse<>("error", "You are not authorized to create a product", null,request.getRequestURI()));
+        }
         return ResponseEntity.ok(new ApiResponse<>("success", "category created", categoryService.createCategory(category), request.getRequestURI()));
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<ApiResponse<Category>> updateCategory(@PathVariable Long id, @RequestBody Category categoryDetails,HttpServletRequest request) {
-            Category updatedCategory = categoryService.updateCategory(id, categoryDetails);
+    public ResponseEntity<ApiResponse<Category>> updateCategory(@RequestAttribute("id") Long userId,@RequestAttribute("role") String role,@PathVariable Long id, @RequestBody Category categoryDetails,HttpServletRequest request) {
+        if(!role.equals(RoleEnum.ADMIN.name())){
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(new ApiResponse<>("error", "You are not authorized to update a product", null,request.getRequestURI()));
+        }
+        Category updatedCategory = categoryService.updateCategory(id, categoryDetails);
             return ResponseEntity.ok(new ApiResponse<>("success", "category updated", updatedCategory, request.getRequestURI()));
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<ApiResponse<String>> deleteCategory(@PathVariable Long id,HttpServletRequest request) {
+    public ResponseEntity<ApiResponse<String>> deleteCategory(@RequestAttribute("id") Long userId,@RequestAttribute("role") String role,@PathVariable Long id,HttpServletRequest request) {
+
+        if(!role.equals(RoleEnum.ADMIN.name())){
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(new ApiResponse<>("error", "You are not authorized to delete a product", null,request.getRequestURI()));
+        }
         categoryService.deleteCategory(id);
         return ResponseEntity.ok(new ApiResponse<>("success", "category deleted", "Category with ID " + id + " has been successfully deleted.", request.getRequestURI()));
     }
